Update 2017-02-19-understanding-symbolic-links.md
I would move the disclaimer about previous knowledge to the top of the file so that it is the first thing that people do if they are uncomfortable.
I don't agree with 'neither a file nor a directory' being bolded, because while it is a useful thing to know, I don't think it's core to the lesson that is trying to be taught here, and overall I felt the article had an overuse of bolding. Similarly, 'text file' on line 119 does not need to be bolded, especially since 'directory' is not bolded earlier in the section.
Line 159 was split after the comma, because the formatting didn't look good in the blog format due to the italics and directory notation being right next to each other before a new line.
Italics were removed from 'could' on line 162, as italics are used for concept or program names, not emphasis throughout the rest of the article.
Line 188 changes are tentative, but it makes it seem like it's a surprise that symlinks work well for config files rather than being a representative example of Linux design.
I feel like the article needs a conclusion but I'm not sure what it would be.